refactor: InputBox and InputStatus components
